+++ src/App.js
@@ -8,13 +8,13 @@ import Login from './components/Authentification/Login/Login.component';
 import InfoCoin from "./components/User/InfoCoin/InfoCoin.component";
 import Profile from "./components/Authentification/Profile/Profile";
 import EditionCoin from "./components/Admin/Adminitration/Edition";
-import Presales from './pages/user/Presales/Presales.component';
+import Presales from './pages/user/Presales/Presales.page';
 import Welcome from "./components/Authentification/Welcome";
-import Administration from './pages/admin/Administration';
-import Submit from './pages/user/Submit/Submit.component';
-import AllTokens from "./pages/user/AllTokens/AllTokens";
+import Administration from './pages/admin/Administration.page';
+import Submit from './pages/user/Submit/Submit.page';
+import AllTokens from "./pages/user/AllTokens/AllTokens.page";
 import { Route, Routes } from 'react-router-dom';
-import ApiConfig from './pages/admin/ApiConfig';
+import ApiConfig from './pages/admin/ApiConfig.page';
 import React from 'react';
 import Home from './Home';
