+++ src/pages/admin/Administration.page.module.scss
@@ -1,0 +1,5 @@
+.administration_fond
+{
+/*     background-color:#050945;
+    min-height: 1000px; */
+}
