+++ src/pages/user/AllTokens/AllTokens.page.js
@@ -1,7 +1,7 @@
 import NavigationUserComponent from '../../../components/Navigation/NavigationUser/NavigationUser.component';
 import AllTokensComponent from '../../../components/User/AllTokens/AllTokens.component';
 import FooterComponent from '../../../components/Navigation/Footer/Footer.component';
-import style from './AllTokens.module.scss';
+import style from './AllTokens.page.module.scss';
 import React from 'react';
